penambahan create customer dan merchant

+++ api/controllers/merchant.controller.js
@@ -25,8 +25,13 @@ exports.createProduct = async (req, res) => {
 exports.updateProduct = async (req, res) => {
     try {
         const product = req.body;
-        await service.updateProduct(product.namaProduct, product.deskripsi, product.harga, product.idProduct, product.stock);
+        let getIdMerchant = await service.getIdMerchantFromProduct(product.idProduct)
 
+        if (product.idMerchant == getIdMerchant) {
+            await service.updateProduct(product.namaProduct, product.deskripsi, product.harga, product.idProduct, product.stock);
+        } else {
+            return res.status(403).json({ code: 403, status: "failed", message: "Anda bukan pemilik produk" })
+        }
         return res.status(200).json({ code: 200, status: "success", message: "Berhasil update data", data: product })
     } catch (err) {
         console.log(err)
@@ -44,9 +49,15 @@ exports.updateProduct = async (req, res) => {
 exports.deleteProduct = async (req, res) => {
     try {
         const product = req.body;
-        const deleteProduct = await service.deleteProduct(product.idProduct);
+        let getIdMerchant = await service.getIdMerchantFromProduct(product.idProduct)
 
-        return res.status(200).json({ code: 200, status: "success", message: "Berhasil delete data", data: deleteProduct })
+        if (product.idMerchant == getIdMerchant) {
+            await service.deleteProduct(product.idProduct);
+        } else {
+            return res.status(403).json({ code: 403, status: "failed", message: "Anda bukan pemilik product" })
+        }
+
+        return res.status(200).json({ code: 200, status: "success", message: "Berhasil delete data", data: product})
     } catch (err) {
         console.log(err)
         return res.status(500).json({
@@ -66,6 +77,25 @@ exports.getListCustomer = async (req, res) => {
         let listCustomer = await service.getListCustomer(merchant);
 
         return res.status(200).json({ code: 200, status: "success", message: "Berhasil melihat list customer", data: listCustomer })
+    } catch (err) {
+        console.log(err)
+        return res.status(500).json({
+            kode: 500,
+            keterangan: "Internal Server Error",
+            data: {
+                kode: '005',
+                keterangan: err.message
+            }
+        })
+    }
+};
+
+exports.createMerchant = async (req, res) => {
+    try {
+        const merchant = req.body;
+        await service.insertMerchant(merchant);
+
+        return res.status(200).json({ code: 200, status: "success", message: "Berhasil menambahkan merchant", data: merchant })
     } catch (err) {
         console.log(err)
         return res.status(500).json({
